Added seemingly redundant save to JenkinsInstanceTVC's delete action to be sure instance deletes are persisted all the way to persistent store.
No longer syncing subviews on subsequent refreshes.

Changed data type and name of JenkinsInstance's "flat" jobs attribute. Added getters and setters for that attribute to convert/unconvert the value into NSDATA

Added SubMenuDelegate Protocol so that submenus can tell the MenuVC to toggle the reveal controller. This closes the menu after saving a JenkinsInstance.

Fixed bug where JenkinsInstanceTVC cancel button was deleting instance always.

added delay before closing menu after saving JenkinsInstance

+++ JenkinsMobile/JenkinsInstanceTableViewController.swift
@@ -11,6 +11,7 @@ import UIKit
 class JenkinsInstanceTableViewController: UITableViewController, UITextFieldDelegate {
     
     var jinstance: JenkinsInstance!
+    var subMenuDelegate: SubMenuDelegate!
     var managedObjectContext: NSManagedObjectContext?
     var showCredentialsFields: Bool?
     var syncMgr: SyncManager?
@@ -376,7 +377,11 @@ class JenkinsInstanceTableViewController: UITableViewController, UITextFieldDele
             self.managedObjectContext?.performBlockAndWait({
                 self.syncMgr?.saveContext(self.managedObjectContext)
             })
+            self.syncMgr?.masterMOC.performBlockAndWait({
+                self.syncMgr?.saveContext(self.syncMgr?.masterMOC)
+            })
             self.activeInstance = false
+            self.saveChanges = true
             self.performSegueWithIdentifier("jenkinsInstanceDoneSegue", sender: self)
         }
         alert.addAction(deleteAction)
@@ -428,10 +433,16 @@ class JenkinsInstanceTableViewController: UITableViewController, UITextFieldDele
             self.jinstance.managedObjectContext?.undoManager?.undo()
         }
         
+        if (segue.identifier == "jenkinsInstanceDoneSegue") {
+            self.subMenuDelegate.revealToggle()
+        }
+        
         if (segue.identifier == "jenkinsInstanceCancelSegue") {
-            let datamgr = DataManager.sharedInstance
-            datamgr.mainMOC.deleteObject(self.jinstance)
-            datamgr.saveContext(datamgr.mainMOC)
+            if (self.jinstance.objectID.temporaryID) {
+                let datamgr = DataManager.sharedInstance
+                datamgr.mainMOC.deleteObject(self.jinstance)
+                datamgr.saveContext(datamgr.mainMOC)
+            }
         }
     }
 }
